+++ LanguageExt.Core/Effects/IO/DSL/IOFoldUntil.cs
@@ -1,0 +1,6 @@
+namespace LanguageExt;
+
+record IOFoldUntil
+{
+    
+}
