+++ LanguageExt.Core/Effects/IO/DSL/IOPure.cs
@@ -1,7 +1,7 @@
 using System;
 using LanguageExt.Traits;
 
-namespace LanguageExt;
+namespace LanguageExt.DSL;
 
 record IOPure<A>(A Value) : InvokeSync<A>
 {
