+++ LanguageExt.Core/Effects/IO/DSL/InvokeSyncIO.cs
@@ -1,23 +1,23 @@
 using System.Threading.Tasks;
 
-namespace LanguageExt;
+namespace LanguageExt.DSL;
 
-abstract record InvokeSyncIO<A> : IO<A>
+public abstract record InvokeSyncIO<A> : IO<A>
 {
     public abstract IO<A> Invoke(EnvIO envIO);
 }
 
-abstract record InvokeAsyncIO<A> : IO<A>
+public abstract record InvokeAsyncIO<A> : IO<A>
 {
     public abstract ValueTask<IO<A>> Invoke(EnvIO envIO);
 }
 
-abstract record InvokeSync<A> : IO<A>
+public abstract record InvokeSync<A> : IO<A>
 {
     public abstract A Invoke(EnvIO envIO);
 }
 
-abstract record InvokeAsync<A> : IO<A>
+public abstract record InvokeAsync<A> : IO<A>
 {
     public abstract ValueTask<A> Invoke(EnvIO envIO);
 }
