+++ LanguageExt.Core/Effects/IO/IO.Prelude.cs
@@ -6,6 +6,7 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using LanguageExt.Common;
+using LanguageExt.DSL;
 using LanguageExt.Traits;
 using LanguageExt.UnsafeValueAccess;
